기능 추가 : 방 생성 기능 추가
