+++ src/main/java/com/sorhive/comprojectserver/room/command/application/controller/RoomController.java
@@ -5,10 +5,7 @@ import com.sorhive.comprojectserver.room.command.application.dto.RoomCreateDto;
 import com.sorhive.comprojectserver.room.command.application.service.RoomService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 
@@ -21,6 +18,7 @@ import javax.validation.Valid;
  * DATE             AUTHOR           NOTE
  * ----------------------------------------------------------------
  * 2022-11-03       부시연           최초 생성
+ * 2022-11-09       부시연           방 생성 추가
  * </pre>
  *
  * @author 부시연(최초 작성자)
@@ -38,8 +36,8 @@ public class RoomController {
 
 
     @PostMapping("room")
-    public ResponseEntity<ResponseDto> createRoom(@Valid @RequestBody RoomCreateDto roomCreateDto) {
+    public ResponseEntity<ResponseDto> createRoom(@RequestHeader String accessToken, @Valid @RequestBody RoomCreateDto roomCreateDto) {
 
-        return ResponseEntity.ok().body(new ResponseDto(HttpStatus.CREATED, "방 생성 성공", roomService.createRoom(roomCreateDto)));
+        return ResponseEntity.ok().body(new ResponseDto(HttpStatus.CREATED, "방 생성 성공", roomService.createRoom(accessToken, roomCreateDto)));
     }
 }
