+++ src/main/java/com/sorhive/comprojectserver/room/command/application/dto/RoomCreateDto.java
@@ -1,11 +1,7 @@
 package com.sorhive.comprojectserver.room.command.application.dto;
 
-import com.sorhive.comprojectserver.member.command.domain.model.member.MemberCode;
-import com.sorhive.comprojectserver.room.command.domain.guestbook.GuestBook;
-import com.sorhive.comprojectserver.room.command.domain.placedfurniture.PlacedFurniture;
-import com.sorhive.comprojectserver.room.command.domain.room.Room;
-
 import java.util.List;
+import java.util.Map;
 
 /**
  * <pre>
@@ -22,23 +18,18 @@ import java.util.List;
  * @version 1(클래스 버전)
  */
 public class RoomCreateDto {
-
     private String floorNumber;
     private String wallNumber;
     private Long memberCode;
-
-    private GuestBookCreateDto guestBook;
-
-    private List<PlacedFurnitureDto> furnitures;
+    private List<Map<String,Object>> furnitures;
 
     public RoomCreateDto() {
     }
 
-    public RoomCreateDto(String floorNumber, String wallNumber, Long memberCode, GuestBookCreateDto guestBook, List<PlacedFurnitureDto> furnitures) {
+    public RoomCreateDto(String floorNumber, String wallNumber, Long memberCode, List<Map<String, Object>> furnitures) {
         this.floorNumber = floorNumber;
         this.wallNumber = wallNumber;
         this.memberCode = memberCode;
-        this.guestBook = guestBook;
         this.furnitures = furnitures;
     }
 
@@ -54,11 +45,7 @@ public class RoomCreateDto {
         return memberCode;
     }
 
-    public GuestBookCreateDto getGuestBook() {
-        return guestBook;
-    }
-
-    public List<PlacedFurnitureDto> getFurnitures() {
+    public List<Map<String, Object>> getFurnitures() {
         return furnitures;
     }
 }
