+++ src/main/java/com/sorhive/comprojectserver/room/command/domain/roomvisit/RoomVisit.java
@@ -1,7 +1,5 @@
 package com.sorhive.comprojectserver.room.command.domain.roomvisit;
 
-import com.sorhive.comprojectserver.room.command.domain.room.Room;
-
 import javax.persistence.*;
 import java.sql.Timestamp;
 
@@ -30,10 +28,6 @@ public class RoomVisit {
 
     @Column(name = "room_visit_time")
     private Timestamp time;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "room_id")
-    private Room room;
 
     @Embedded
     private RoomVisitor roomVisitor;
